+++ app/src/main/java/org/autojs/autoxjs/ui/main/drawer/DrawerPage.kt
@@ -1774,6 +1774,30 @@ fun detailsDialog(context: Context){
         .item(
             R.id.modification_detail,
             R.drawable.ic_edit_black_48dp,
+            "<=== 65816 ===>"
+        )
+        .item(
+            R.id.modification_detail,
+            R.drawable.ic_ali_log,
+            "添加: 设置 button 渐变背景\n"+
+            "> <button id=\"btn\" w=\"88\" h=\"88\" gradient=\"shape=oval|colors=#ff00ff,#584EF0|ori=bottom_top|type=linear\">\n" +
+            "> ui.btn.setBackgroundGradient(\"shape=rect|corner=88\");\n" +
+            "> 参数：shape: rect(方形-默认)、line(线)、ring(圆环)、oval(椭圆)\n" +
+            "> colors: 渐变颜色数组\n" +
+            "> ori: 渐变方向 top_bottom、bottom_top、left_right、right_left、tl_br、br_tl、tr_bl、bl_tr\n" +
+            "> type: 渐变类型 linear(线性-默认) radial(辐射) sweep(扫描)\n" +
+            "> center: 渐变中心 0.5,0.5  默认(x:0.5, y:0.5)\n" +
+            "> corner: 圆角 默认16\n" +
+            "> 有些可能不符合预期，暂时不深入研究 *.*\n\n"+
+            "添加: 设置 radio 色调\n"+
+            "> <radio id=\"radio\" tint=\"#ff0000|#00ff00\" />\n" +
+            "> ui.radio.setTint(\"#00ff00|#ff0000\")\n" +
+            "> 注：未选中|选中 （只有一个颜色则一样）\n\n"+
+            "修改：button、input、spinner、radio、text、toolbar 控件为 androidx 的控件"
+        )
+        .item(
+            R.id.modification_detail,
+            R.drawable.ic_edit_black_48dp,
             "<=== 65815 ===>"
         )
         .item(