+++ autojs/src/main/java/com/stardust/autojs/core/ui/inflater/util/Colors.java
@@ -4,8 +4,6 @@ import android.content.Context;
 import android.content.res.ColorStateList;
 import android.content.res.Resources;
 import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
-import android.graphics.drawable.StateListDrawable;
 import android.view.View;
 
 import com.stardust.app.GlobalAppContext;