+++ autojs/src/main/java/com/stardust/autojs/core/ui/widget/JsEditText.java
@@ -2,18 +2,14 @@ package com.stardust.autojs.core.ui.widget;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.os.Build;
 import android.util.AttributeSet;
-import android.widget.EditText;
-
-import androidx.annotation.RequiresApi;
 
 /**
  * Created by Stardust on 2017/5/15.
  */
 
 @SuppressLint("AppCompatCustomView")
-public class JsEditText extends EditText {
+public class JsEditText extends androidx.appcompat.widget.AppCompatEditText {
     public JsEditText(Context context) {
         super(context);
     }
@@ -24,12 +20,13 @@ public class JsEditText extends EditText {
 
     public JsEditText(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
+
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-    public JsEditText(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
-        super(context, attrs, defStyleAttr, defStyleRes);
-    }
+//    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+//    public JsEditText(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+//        super(context, attrs, defStyleAttr, defStyleRes);
+//    }
 
     public String text() {
         return getText().toString();