+++ autojs/src/main/java/com/stardust/autojs/core/ui/widget/JsGridView.java
@@ -3,7 +3,6 @@ package com.stardust.autojs.core.ui.widget;
 import android.content.Context;
 
 import androidx.recyclerview.widget.GridLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import com.stardust.autojs.runtime.ScriptRuntime;
 
@@ -22,7 +21,7 @@ public class JsGridView extends JsListView {
         super.init();
         setLayoutManager(new GridLayoutManager(getContext(), 1){
             @Override
-            public void onLayoutChildren(RecyclerView.Recycler recycler, RecyclerView.State state) {
+            public void onLayoutChildren(Recycler recycler, State state) {
                 try {
                     super.onLayoutChildren(recycler, state);
                 } catch (IndexOutOfBoundsException e) {